requirements.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
docker-pycreds==0.4.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.12.0
fonttools==4.37.1
fsspec==2023.6.0
gdown==4.7.1
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.16.4
idna==3.3
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///croot/ipython_1668088122043/work
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
joblib==1.1.0
jupyter_client @ file:///opt/conda/conda-bld/jupyter_client_1662504365333/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847906199/work
numpy==1.23.2
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
psutil==5.9.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.2.1
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
requests==2.28.1
safetensors==0.3.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
soupsieve==2.4.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
threadpoolctl==3.1.0
timm==0.9.2
torch==1.12.0+cu113
torchaudio==0.12.0+rocm5.1.1
torchvision==0.13.0+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
yacs==0.1.8
